refactor: web3-shared utils
