+++ packages/web3-shared/src/utils/index.ts
@@ -1,234 +1,5 @@
-import BigNumber from 'bignumber.js'
-import { getEnumAsArray, safeUnreachable } from '@dimensiondev/maskbook-shared'
-import CHAINS from '../assets/chains.json'
-import { TOKEN_CONSTANTS } from '../constants'
-import { constantOfChain } from '../hooks'
-import {
-    Asset,
-    ChainId,
-    CurrencyType,
-    ERC1155TokenAssetDetailed,
-    ERC20TokenDetailed,
-    ERC721TokenAssetDetailed,
-    EthereumTokenType,
-    NativeTokenDetailed,
-    NetworkType,
-    Web3Constants,
-} from '../types'
-
-export function isSameAddress(addrA: string, addrB: string) {
-    return addrA.toLowerCase() === addrB.toLowerCase()
-}
-
-export function currySameAddress(base: string) {
-    return (target: string | { address: string }) => {
-        if (typeof target === 'string') {
-            return isSameAddress(base, target)
-        } else if (typeof target === 'object' && typeof target.address === 'string') {
-            return isSameAddress(base, target.address)
-        }
-        throw new Error('Unsupported `target` address format')
-    }
-}
-
-export const isDAI = currySameAddress(constantOfChain(TOKEN_CONSTANTS, ChainId.Mainnet).DAI_ADDRESS)
-
-export const isOKB = currySameAddress(constantOfChain(TOKEN_CONSTANTS, ChainId.Mainnet).OKB_ADDRESS)
-
-export const isNative = currySameAddress(constantOfChain(TOKEN_CONSTANTS, ChainId.Mainnet).NATIVE_TOKEN_ADDRESS)
-
-export function addGasMargin(value: BigNumber.Value, scale = 3000) {
-    return new BigNumber(value).multipliedBy(new BigNumber(10000).plus(scale)).dividedToIntegerBy(10000)
-}
-
-//#region constants
-
-/**
- * @deprecated Use constantOfChain from @dimensiondev/web3-shared package
- *
- * Before: `getConstant(T, "a", ChainId.Mainnet)`
- *
- * After: `constantOfChain(T, ChainId.Mainnet).a`
- */
-export function getConstant<T extends Web3Constants, K extends keyof T>(
-    constants: T,
-    key: K,
-    chainId = ChainId.Mainnet,
-): T[K][ChainId.Mainnet] {
-    return constants[key][chainId]
-}
-//#endregion
-
-//#region chain detailed
-export function getChainDetailed(chainId: ChainId = ChainId.Mainnet) {
-    return CHAINS.find((x) => x.chainId === chainId)
-}
-
-// Learn more: https://github.com/ChainAgnostic/CAIPs/blob/master/CAIPs/caip-2.md
-export function getChainDetailedCAIP(chainId: ChainId = ChainId.Mainnet) {
-    const chainDetailed = getChainDetailed(chainId)
-    if (!chainDetailed) return
-    return {
-        chainId: `0x${chainDetailed.chainId.toString(16)}`,
-        chainName: chainDetailed.name,
-        nativeCurrency: chainDetailed.nativeCurrency,
-        rpcUrls: chainDetailed.rpc,
-        blockExplorerUrls: [
-            chainDetailed.explorers && chainDetailed.explorers.length > 0 && chainDetailed.explorers[0].url
-                ? chainDetailed.explorers[0].url
-                : chainDetailed.infoURL,
-        ],
-    }
-}
-
-export function getChainName(chainId: ChainId) {
-    const chainDetailed = getChainDetailed(chainId)
-    return chainDetailed?.name ?? 'Unknown'
-}
-
-export function getChainFullName(chainId: ChainId) {
-    const chainDetailed = getChainDetailed(chainId)
-    return chainDetailed?.fullName ?? 'Unknown'
-}
-
-export function getChainIdFromName(name: string) {
-    const chainDetailed = CHAINS.find((x) =>
-        [x.chain, x.network, x.name, x.shortName, x.fullName].map((y) => y.toLowerCase()).includes(name.toLowerCase()),
-    )
-    return chainDetailed?.chainId as ChainId | undefined
-}
-
-export function getChainIdFromNetworkType(networkType: NetworkType) {
-    switch (networkType) {
-        case NetworkType.Ethereum:
-            return ChainId.Mainnet
-        case NetworkType.Binance:
-            return ChainId.BSC
-        case NetworkType.Polygon:
-            return ChainId.Matic
-        default:
-            safeUnreachable(networkType)
-            return ChainId.Mainnet
-    }
-}
-
-export function getNetworkTypeFromChainId(chainId: ChainId) {
-    const chainDetailed = getChainDetailed(chainId)
-    switch (chainDetailed?.chain) {
-        case 'ETH':
-            return NetworkType.Ethereum
-        case 'BSC':
-            return NetworkType.Binance
-        case 'Matic':
-            return NetworkType.Polygon
-        default:
-            throw new Error('Unknown chain id.')
-    }
-}
-//#endregion
-
-//#region tokens
-export function createNativeToken(chainId: ChainId): NativeTokenDetailed {
-    const chainDetailed = getChainDetailed(chainId)
-    if (!chainDetailed) throw new Error('Unknown chain id.')
-    return {
-        type: EthereumTokenType.Native,
-        chainId,
-        address: getConstant(TOKEN_CONSTANTS, 'NATIVE_TOKEN_ADDRESS'),
-        ...chainDetailed.nativeCurrency,
-    }
-}
-
-export function createERC20Token(
-    chainId: ChainId,
-    address: string,
-    decimals: number,
-    name: string,
-    symbol: string,
-): ERC20TokenDetailed {
-    return {
-        type: EthereumTokenType.ERC20,
-        chainId,
-        address,
-        decimals,
-        name,
-        symbol,
-    }
-}
-
-export function createERC721Token(
-    chainId: ChainId,
-    tokenId: string,
-    address: string,
-    name: string,
-    symbol: string,
-    baseURI?: string,
-    tokenURI?: string,
-    asset?: ERC721TokenAssetDetailed['asset'],
-): ERC721TokenAssetDetailed {
-    return {
-        type: EthereumTokenType.ERC721,
-        chainId,
-        tokenId,
-        address,
-        name,
-        symbol,
-        baseURI,
-        tokenURI,
-        asset,
-    }
-}
-
-export function createERC1155Token(
-    chainId: ChainId,
-    tokenId: string,
-    address: string,
-    name: string,
-    uri?: string,
-    asset?: ERC1155TokenAssetDetailed['asset'],
-): ERC1155TokenAssetDetailed {
-    return {
-        type: EthereumTokenType.ERC1155,
-        chainId,
-        tokenId,
-        address,
-        name,
-        uri,
-        asset,
-    }
-}
-
-export function createERC20Tokens(
-    key: keyof typeof TOKEN_CONSTANTS,
-    name: string | ((chainId: ChainId) => string),
-    symbol: string | ((chainId: ChainId) => string),
-    decimals: number | ((chainId: ChainId) => number),
-) {
-    return getEnumAsArray(ChainId).reduce((accumulator, { value: chainId }) => {
-        const evaludator: <T>(f: T | ((chainId: ChainId) => T)) => T = (f) =>
-            typeof f === 'function' ? (f as any)(chainId) : f
-
-        accumulator[chainId] = {
-            type: EthereumTokenType.ERC20,
-            chainId,
-            address: getConstant(TOKEN_CONSTANTS, key, chainId) as string,
-            name: evaludator(name),
-            symbol: evaludator(symbol),
-            decimals: evaludator(decimals),
-        }
-        return accumulator
-    }, {} as { [chainId in ChainId]: ERC20TokenDetailed })
-}
-//#endregion
-
-//#region web3
-import type Web3 from 'web3'
-import type { AbiOutput } from 'web3-utils'
-
-export function decodeOutputString(web3: Web3, abis: AbiOutput[], output: string) {
-    if (abis.length === 1) return web3.eth.abi.decodeParameter(abis[0], output)
-    if (abis.length > 1) return web3.eth.abi.decodeParameters(abis, output)
-    return
-}
-
-export const getTokenUSDValue = (token: Asset) => (token.value ? Number.parseFloat(token.value[CurrencyType.USD]) : 0)
+export * from './address'
+export * from './chain'
+export * from './constant'
+export * from './token'
+export * from './web3'
