+++ packages/web3-shared/src/utils/web3.ts
@@ -1,0 +1,16 @@
+import type Web3 from 'web3'
+import type { AbiOutput } from 'web3-utils'
+import BigNumber from 'bignumber.js'
+import { Asset, CurrencyType } from '../types'
+
+export function decodeOutputString(web3: Web3, abis: AbiOutput[], output: string) {
+    if (abis.length === 1) return web3.eth.abi.decodeParameter(abis[0], output)
+    if (abis.length > 1) return web3.eth.abi.decodeParameters(abis, output)
+    return
+}
+
+export function addGasMargin(value: BigNumber.Value, scale = 3000) {
+    return new BigNumber(value).multipliedBy(new BigNumber(10000).plus(scale)).dividedToIntegerBy(10000)
+}
+
+export const getTokenUSDValue = (token: Asset) => (token.value ? Number.parseFloat(token.value[CurrencyType.USD]) : 0)
